+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.56.0] - 2024-10-21
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Semi-transparent black overlay to landing page screens.
+- [Marek Kurańda](https://github.com/mjkuranda): Icons for buttons: `log in`, `log out`, `manage`, `search`, `best search` and `clear`.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): `Yummy` to `Co na dziś?`.
+
 ## [1.55.0] - 2024-10-21
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Activation page at `/users/activate/[activationCode]`.